pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strux"
version = "0.1.0"
description = "Ontology-oriented student career decision support prototype"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
  "fastapi>=0.135.2,<1.0",
  "networkx>=3.5,<4.0",
  "openai>=2.6.1,<3.0",
  "openpyxl>=3.1.5,<4.0",
  "pandas>=3.0.1,<4.0",
  "pydantic>=2.12.5,<3.0",
  "pydantic-settings>=2.13.1,<3.0",
  "PyYAML>=6.0.3,<7.0",
  "uvicorn>=0.42.0,<1.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1,<1.0",
  "pytest>=9.0.2,<10.0",
  "ruff>=0.15.8,<1.0",
]

[project.scripts]
strux = "strux.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py314"

[tool.pytest.ini_options]
testpaths = ["tests"]
